backend/apiBnshop/requirements.txt
asgiref==3.5.2
distlib==0.3.5
dj-database-url==1.0.0
Django==4.0.6
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
gunicorn==20.1.0
Markdown==3.4.1
platformdirs==2.5.2
psycopg==3.0.16
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.3
whitenoise==6.2.0
facebook-sdk==3.1.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
requests==2.23.0
requests-oauthlib==1.3.0
text-unidecode==1.3